Move files from `sources.remote` into `remote`
