+++ data/src/main/java/com/segunfamisa/zeitung/data/di/DataModule.kt
@@ -7,12 +7,12 @@ import com.segunfamisa.zeitung.data.headlines.RemoteHeadlinesSource
 import com.segunfamisa.zeitung.data.newssources.NewsSourcesDataSource
 import com.segunfamisa.zeitung.data.newssources.NewsSourcesRepositoryImpl
 import com.segunfamisa.zeitung.data.newssources.RemoteNewsSourcesDataSource
-import com.segunfamisa.zeitung.data.sources.remote.ApiKeyProvider
-import com.segunfamisa.zeitung.data.sources.remote.ApiKeyProviderImpl
-import com.segunfamisa.zeitung.data.sources.remote.ApiService
-import com.segunfamisa.zeitung.data.sources.remote.ApiServiceCreator
-import com.segunfamisa.zeitung.data.sources.remote.UrlProvider
-import com.segunfamisa.zeitung.data.sources.remote.UrlProviderImpl
+import com.segunfamisa.zeitung.data.remote.ApiKeyProvider
+import com.segunfamisa.zeitung.data.remote.ApiKeyProviderImpl
+import com.segunfamisa.zeitung.data.remote.ApiService
+import com.segunfamisa.zeitung.data.remote.ApiServiceCreator
+import com.segunfamisa.zeitung.data.remote.UrlProvider
+import com.segunfamisa.zeitung.data.remote.UrlProviderImpl
 import com.segunfamisa.zeitung.domain.headlines.HeadlinesRepository
 import com.segunfamisa.zeitung.domain.newssources.NewsSourcesRepository
 import dagger.Binds
