+++ data/src/main/java/com/segunfamisa/zeitung/data/newssources/RemoteNewsSourcesDataSource.kt
@@ -2,7 +2,7 @@ package com.segunfamisa.zeitung.data.newssources
 
 import arrow.core.Either
 import com.segunfamisa.zeitung.core.entities.Source
-import com.segunfamisa.zeitung.data.sources.remote.ApiService
+import com.segunfamisa.zeitung.data.remote.ApiService
 import com.segunfamisa.zeitung.domain.common.Error
 import javax.inject.Inject
