+++ data/src/test/java/com/segunfamisa/zeitung/data/newssources/RemoteNewsSourcesDataSourceTest.kt
@@ -3,9 +3,9 @@ package com.segunfamisa.zeitung.data.newssources
 import arrow.core.orNull
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.whenever
-import com.segunfamisa.zeitung.data.sources.remote.ApiService
+import com.segunfamisa.zeitung.data.remote.ApiService
 import com.segunfamisa.zeitung.data.sources.remote.TestDataGenerator
-import com.segunfamisa.zeitung.data.sources.remote.entities.SourcesResponse
+import com.segunfamisa.zeitung.data.remote.entities.SourcesResponse
 import kotlinx.coroutines.async
 import kotlinx.coroutines.runBlocking
 import org.junit.Assert.assertEquals
