+++ data/src/test/java/com/segunfamisa/zeitung/data/sources/remote/TestDataGenerator.kt
@@ -1,8 +1,8 @@
 package com.segunfamisa.zeitung.data.sources.remote
 
-import com.segunfamisa.zeitung.data.sources.remote.entities.Article
-import com.segunfamisa.zeitung.data.sources.remote.entities.Source
-import com.segunfamisa.zeitung.data.sources.remote.entities.SourceMinimal
+import com.segunfamisa.zeitung.data.remote.entities.Article
+import com.segunfamisa.zeitung.data.remote.entities.Source
+import com.segunfamisa.zeitung.data.remote.entities.SourceMinimal
 import java.util.Date
 
 internal object TestDataGenerator {
